Change ship and gameboard to return true or false when attacking a ship
Before this commit, the ship hit function was returning a message reporting
if the hit landed, hit the same location twice, and if it was a successful. Now
the hit function will only return true or false depending on the hit. The same
thing was changed with the receieveAttack function in the gameboard.

Tests were changed to reflect these changes